Adds an initial working version of the figures counter
The current implementation accepts only input from the standard input and offers no user interface. It scans the input line by line using at most two rows of the input and holds a map of the equivalent figure IDs from which we find the unique IDs in the end. Includes a test that covers one of the more tricky cases - a circular shape.